Split in different tasks.
