+++ README.md
@@ -124,6 +124,10 @@ Returns an object with all contracts from a deployment and their addresses.
 | V3 Vault                                               | [`20241204-v3-vault`](./v3/tasks/20241204-v3-vault)                                                     |
 | V3 Weighted Pool                                       | [`20241205-v3-weighted-pool`](./v3/tasks/20241205-v3-weighted-pool)                                     |
 | V3 Stable Pool                                         | [`20241205-v3-stable-pool`](./v3/tasks/20241205-v3-stable-pool)                                         |
+| V3 Router                                              | [`20241205-v3-router`](./v3/tasks/20241205-v3-router)                                                   |
+| V3 Batch Router                                        | [`20241205-v3-batch-router`](./v3/tasks/20241205-v3-batch-router)                                       |
+| V3 Composite Liquidity Router                          | [`20241205-v3-composite-liquidity-router`](./v3/tasks/20241205-v3-composite-liquidity-router)           |
+| V3 Buffer Router                                       | [`20241205-v3-buffer-router`](./v3/tasks/20241205-v3-buffer-router)                                     |
 
 ## Scripts
