+++ v3/tasks/20241205-v3-batch-router/readme.md
@@ -1,0 +1,11 @@
+# V3 Batch Router
+
+Batch Router deployment for V3.
+Contains `BatchRouter` for complex multi step swaps (supports single token add / remove types and buffer wrap / unwrap).
+
+## Useful Files
+
+- [Code](https://github.com/balancer/balancer-v3-monorepo/commit/74d7068fb21565741427cdabfa4f1b539a4bddaa).
+- [Sepolia testnet addresses](./output/sepolia.json)
+- [`BatchRouter` artifact](./artifact/BatchRouter.json)
+
