+++ v3/tasks/20241205-v3-buffer-router/readme.md
@@ -1,0 +1,11 @@
+# V3 Buffer Router
+
+Buffer Router set deployment for V3.
+Contains `BufferRouter` to initialize and manage liquidity positions for buffers.
+
+## Useful Files
+
+- [Code](https://github.com/balancer/balancer-v3-monorepo/commit/74d7068fb21565741427cdabfa4f1b539a4bddaa).
+- [Sepolia testnet addresses](./output/sepolia.json)
+- [`BufferRouter` artifact](./artifact/BufferRouter.json)
+
