+++ v3/tasks/20241205-v3-composite-liquidity-router/readme.md
@@ -1,0 +1,11 @@
+# V3 Composite Liquidity Router
+
+Composite Liquidity Router deployment for V3.
+Contains `CompositeLiquidityRouter` for complex liquidity operations involving pools with ERC4626 wrappers and nested pools.
+
+## Useful Files
+
+- [Code](https://github.com/balancer/balancer-v3-monorepo/commit/74d7068fb21565741427cdabfa4f1b539a4bddaa).
+- [Sepolia testnet addresses](./output/sepolia.json)
+- [`CompositeLiquidityRouter` artifact](./artifact/CompositeLiquidityRouter.json)
+
