+++ v3/tasks/20241205-v3-router/readme.md
@@ -1,0 +1,11 @@
+# V3 Router
+
+Base Router deployment for V3.
+Contains `Router` for basic, single step operation (pool initialization, add, remove, swap).
+
+## Useful Files
+
+- [Code](https://github.com/balancer/balancer-v3-monorepo/commit/74d7068fb21565741427cdabfa4f1b539a4bddaa).
+- [Sepolia testnet addresses](./output/sepolia.json)
+- [`Router` artifact](./artifact/Router.json)
+
